# moviemix/backend/src/search/search.controller.ts
import { Controller, Get, Query } from '@nestjs/common';
import { SearchService } from './search.service';

@Controller('search')
export class SearchController {
  constructor(private readonly searchService: SearchService) {}

  @Get()
  async search(
    @Query('q') q: string,
    @Query('type') type?: string,
    @Query('page') page?: number,
    @Query('limit') limit?: number,
  ) {
    return this.searchService.search(q, type, page, limit);
  }

  @Get('suggestions')
  async suggestions(@Query('q') q: string) {
    return this.searchService.suggestions(q);
  }
}